commit 0.57 Added notification for developer if task is bought and updatede task process functionality

+++ src/main/java/org/example/kihelp_back/history/repository/HistoryRepository.java
@@ -10,5 +10,5 @@ public interface HistoryRepository extends JpaRepository<History, Long> {
     @Transactional(readOnly = true)
     List<History> findAllByUserTelegramId(String telegramId);
     @Transactional(readOnly = true)
-    List<History> findAllByTaskId(Integer taskId);
+    List<History> findAllByTaskId(Long taskId);
 }
